Added MWFeedReader and replaced the old feed parser with MWFeedReader for the news section
First modification to include the MWFeedReader.  The news section has
been modified to use the new feed parser.  This adds support for atom
feeds.  Lots of code to still cleanup but it is a start.  Still
supports RSS as well.  My intent is to have it replace the FeedLoader
class in all the sections of the application. MWFeedReader is an open
source project that is actively maintain and should support better
parsing for Communique.